requirements.txt
apricot-select==0.6.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
joblib==1.2.0
kiwisolver==1.4.4
kmedoids==0.4.3
llvmlite==0.40.0
matplotlib==3.7.1
nose==1.3.7
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
scikit-learn-extra==0.3.0
banditpam
